feat: bulk upload UI
